pyproject.toml
[project]
name = "obsidian-infra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "google-api-python-client>=2.190.0",
  "google-api-python-client-stubs[drive]>=1.32.0",
  "google-auth-httplib2>=0.3.0",
  "google-auth-oauthlib>=1.3.0",
  "python-dotenv>=1.2.1",
  "requests>=2.31.0",
]
